able to update the DB correctly
